+++ sdk/loadtesting/playwright/src/core/playwrightService.ts
@@ -18,10 +18,10 @@ import {
   validatePlaywrightVersion,
   validateServiceUrl,
   exitWithFailureMessage,
-  getPackageVersion,
   getPlaywrightVersion,
   getVersionInfo,
   throwErrorWithFailureMessage,
+  getPackageVersion,
 } from "../utils/utils.js";
 import { ServiceErrorMessageConstants } from "../common/messages.js";
 import type { PlaywrightTestConfig } from "@playwright/test";